The role table has been removed. Role management will continue via Role Type enum.

+++ Flowey.BUSINESS/Features/Projects/Commands/AddProjectWithCreatorCommand.cs
@@ -52,7 +52,7 @@ namespace Flowey.BUSINESS.Features.Projects.Commands
                     new ProjectUserRole
                     {
                         UserId = _currentUserService.GetUserId().Value,
-                        RoleId = (int)RoleType.Admin
+                        RoleId = RoleType.Admin
                     }
                 }
             };
